+++ src/Sentry.Maui/Internal/SentryMauiScreenshotProcessor.cs
@@ -18,6 +18,12 @@ internal class SentryMauiScreenshotProcessor : ISentryEventProcessorWithHint
 
     public SentryEvent? Process(SentryEvent @event, SentryHint hint)
     {
+        // Call back before taking the screenshot if the callback is not null
+        if (_options.BeforeCaptureScreenshotInternal is not null)
+        {
+            _options.BeforeCaptureScreenshotInternal.Invoke(@event, hint);
+        }
+
         hint.Attachments.Add(new ScreenshotAttachment(_options));
         return @event;
     }